set constant in runtime instead
