+++ chart/templates/services/search/_container.tpl
@@ -37,8 +37,6 @@
     value: {{ .Values.duckDBIndex.cacheDirectory | quote }}
   - name: DUCKDB_INDEX_EXTENSIONS_DIRECTORY
     value: "/tmp/duckdb-extensions"
-  - name: HF_HUB_ENABLE_HF_TRANSFER
-    value: 1
   volumeMounts:
   {{ include "volumeMountDuckDBIndexRW" . | nindent 2 }}
   securityContext:
