+++ services/search/Dockerfile
@@ -12,8 +12,7 @@ ENV PYTHONFAULTHANDLER=1 \
     # Versions:
     POETRY_VERSION=1.7.1 \
     POETRY_VIRTUALENVS_IN_PROJECT=true \
-    PATH="$PATH:/root/.local/bin" \
-    HF_HUB_ENABLE_HF_TRANSFER=1
+    PATH="$PATH:/root/.local/bin" 
 
 # System deps:
 RUN apt-get update \
