+++ services/search/Makefile
@@ -3,7 +3,6 @@ export COMPOSE_PROJECT_NAME := search
 export MONGO_PORT := 27033
 export CACHE_MONGO_URL := mongodb://localhost:${MONGO_PORT}
 export QUEUE_MONGO_URL := mongodb://localhost:${MONGO_PORT}
-export HF_HUB_ENABLE_HF_TRANSFER := 1
 
 # makefile variables
 DOCKER_COMPOSE := ../../tools/docker-compose-mongo.yml
