+++ services/worker/src/worker/job_runners/config/parquet_and_info.py
@@ -71,18 +71,17 @@ from libcommon.exceptions import (
 from libcommon.parquet_utils import PART_SUFFIX, PARTIAL_PREFIX
 from libcommon.queue import lock
 from libcommon.simple_cache import get_previous_step_or_raise
+from libcommon.utils import HF_HUB_HTTP_ERROR_RETRY_SLEEPS, retry
 from tqdm.contrib.concurrent import thread_map
 
 from worker.config import AppConfig, ParquetAndInfoConfig
 from worker.dtos import CompleteJobResult, ConfigParquetAndInfoResponse
 from worker.job_runners.config.config_job_runner import ConfigJobRunnerWithDatasetsCache
 from worker.utils import (
-    HF_HUB_HTTP_ERROR_RETRY_SLEEPS,
     LOCK_GIT_BRANCH_RETRY_SLEEPS,
     create_branch,
     hf_hub_url,
     resolve_trust_remote_code,
-    retry,
 )
 
 DATASET_TYPE = "dataset"
